[AN] chore: 데이터 레이어 패키지 분리 #111
