--- android/app/src/main/java/com/onair/hearit/data/AuthRemoteDataSource.kt
@@ -1,5 +1,0 @@
-package com.onair.hearit.data
-
-interface AuthRemoteDataSource {
-    suspend fun kakaoLogin(kakaoLoginRequest: KakaoLoginRequest): Result<KakaoLoginResponse>
-}
