+++ android/app/src/main/java/com/onair/hearit/data/api/AuthService.kt
@@ -1,5 +1,7 @@
-package com.onair.hearit.data
+package com.onair.hearit.data.api
 
+import com.onair.hearit.data.dto.KakaoLoginRequest
+import com.onair.hearit.data.dto.KakaoLoginResponse
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.POST
