+++ android/app/src/main/java/com/onair/hearit/data/api/HearitService.kt
@@ -1,5 +1,8 @@
-package com.onair.hearit.data
+package com.onair.hearit.data.api
 
+import com.onair.hearit.data.dto.RandomHearitResponse
+import com.onair.hearit.data.dto.RecommendHearitResponse
+import com.onair.hearit.data.dto.SearchHearitResponse
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Query
