+++ android/app/src/main/java/com/onair/hearit/data/datasource/AuthRemoteDataSource.kt
@@ -1,0 +1,8 @@
+package com.onair.hearit.data.datasource
+
+import com.onair.hearit.data.dto.KakaoLoginRequest
+import com.onair.hearit.data.dto.KakaoLoginResponse
+
+interface AuthRemoteDataSource {
+    suspend fun kakaoLogin(kakaoLoginRequest: KakaoLoginRequest): Result<KakaoLoginResponse>
+}
