+++ android/app/src/main/java/com/onair/hearit/data/datasource/MediaFileRemoteDataSource.kt
@@ -1,5 +1,7 @@
-package com.onair.hearit.data
+package com.onair.hearit.data.datasource
 
+import com.onair.hearit.data.dto.ScriptUrlResponse
+import com.onair.hearit.data.dto.ShortAudioUrlResponse
 import okhttp3.ResponseBody
 
 interface MediaFileRemoteDataSource {
