+++ android/app/src/main/java/com/onair/hearit/data/dto/RecommendHearitResponse.kt
@@ -1,25 +1,28 @@
-package com.onair.hearit.data
+package com.onair.hearit.data.dto
 
-import com.onair.hearit.domain.SearchedHearitItem
+import com.onair.hearit.domain.RecommendHearitItem
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class SearchHearitResponse(
+data class RecommendHearitResponse(
+    @SerialName("createdAt")
+    val createdAt: String,
     @SerialName("id")
     val id: Long,
     @SerialName("playTime")
     val playTime: Int,
+    @SerialName("source")
+    val source: String,
     @SerialName("summary")
     val summary: String,
     @SerialName("title")
     val title: String,
 )
 
-fun SearchHearitResponse.toDomain(): SearchedHearitItem =
-    SearchedHearitItem(
+fun RecommendHearitResponse.toDomain(): RecommendHearitItem =
+    RecommendHearitItem(
         id = this.id,
         title = this.title,
-        playTime = this.playTime,
-        summary = this.summary,
+        desc = this.summary,
     )
