+++ android/app/src/main/java/com/onair/hearit/data/dto/ShortAudioUrlResponse.kt
@@ -1,4 +1,4 @@
-package com.onair.hearit.data
+package com.onair.hearit.data.dto
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
